pyproject.toml
[project]
name = "imagegen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "httpx>=0.28.1",
    "openai>=1.76.0",
    "pillow>=11.2.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
